2005-10-05 15:05 UTC+0100 Ryszard Glab <rglab@imid.med.pl>    * source/common/expropt2.c       * fixed bug in ASC() optimization (when character code > 127)

+++ harbour/source/common/expropt2.c
@@ -1231,7 +1231,7 @@ BOOL hb_compExprReduceASC( HB_EXPR_PTR pSelf, HB_MACRO_DECL )
 	
    if( pArg->ExprType == HB_ET_STRING )
    {
-   	HB_EXPR_PTR pExpr = hb_compExprNewLong( pArg->value.asString.string[0] );
+   	HB_EXPR_PTR pExpr = hb_compExprNewLong( (BYTE)pArg->value.asString.string[0] );
 
      	hb_compExprFree( pParms, HB_MACRO_PARAM );
      	hb_compExprFree( pSelf->value.asFunCall.pFunName, HB_MACRO_PARAM );
